api/requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
Flask-Limiter==1.0.1
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
limits==1.3
MarkupSafe==1.0
pysha3==1.0.2
python-rocksdb==0.6.9
pytz==2018.5
six==1.11.0
uWSGI==2.0.17.1
Werkzeug==0.14.1
